Fix: dependency #26
