+++ backend/nestjs-app/src/game/game.controller.ts
@@ -1,10 +1,14 @@
 import { Body, Controller, Post, UseGuards, Request } from '@nestjs/common';
 import JwtTwoFactorGuard from 'src/auth/jwt/jwt-two-factor.gaurd';
 import { GameService } from './game.service';
+import { GameGateway } from './game.gateway';
 
 @Controller('game')
 export class GameController {
-  constructor(private readonly gameService: GameService) {}
+  constructor(
+    private readonly gameService: GameService,
+    private readonly gameGateway: GameGateway,
+  ) {}
 
   @Post('battle/offer')
   @UseGuards(JwtTwoFactorGuard)
@@ -13,6 +17,6 @@ export class GameController {
     @Body('id') user_id: string,
     @Body('nickname') nickname: string,
   ): Promise<boolean> {
-    return await this.gameService.offerGame(user_id, nickname);
+    return await this.gameGateway.offerGame(user_id, nickname);
   }
 }
